Add eslint.
To get the javascript in shape, we can add eslint to
the project, but we need to have a set of files we want to
target. I think it might be good to move the webpack
code to a separate folder and then we lint stuff as we move things there.

Added 'dev' target to package.json.  It runs webpack with watch mode
and -d to put webpack in debug mode so it generates sourcemaps.